refractored the shoppingcart.
