+++ src/js/product.js
@@ -9,5 +9,3 @@ const productId = getParam("product");
 
 const product = new ProductDetails(productId, dataSource);
 product.init();
-
-
