+++ src/product-listing/index.html
@@ -11,7 +11,7 @@
     <header class="divider" id="main-header"></header>
     <main class="divider">
       <section class="products">
-        <h2 class="title">Top Products: </h2>
+        <h2 class="title">Top Products:</h2>
         <ul class="product-list"></ul>
       </section>
     </main>
